% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_entso_e.R
\name{download_entso_e}
\alias{download_entso_e}
\title{Download data from ENTSO-E Transparency Platform}
\usage{
download_entso_e(
  data_item = NULL,
  path = paste0(get_credentials()$data_path, "entsoe_TP/"),
  userpwd = get_credentials()$entso_e
)
}
\arguments{
\item{data_item}{Name of data item. If \code{NULL}, a list of options is returned and saved to \code{path})}

\item{path}{Path to folder for entso-e data}

\item{userpwd}{Credentials for the entso-e transparency platform}
}
\description{
Download or update a complete copy of files available on the entso-e transparency platform
for a given data item.
}
\details{
This function downloads all data matching \code{data_item} from the entso-e transparency platform,
that are not already present in \code{path}. Files that are present but have not been updated since the end
of the time period that they correspond to will be updated. E.g. incomplete files for the present and/or past month
are typically available, and this function will update the local copy of them.

Further documentation may be found at \link{https://transparency.entsoe.eu/}.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
