% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{get_credentials}
\alias{get_credentials}
\title{Functions to get FTP/API credentials}
\usage{
get_credentials()
}
\value{
A \code{list} of credentials.
}
\description{
Get credentials from file created by \code{make_credentials}.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
