% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{make_credentials}
\alias{make_credentials}
\title{Function to set FTP/API credentials and set path to data}
\usage{
make_credentials(
  n2ex = NULL,
  entso_e = NULL,
  data_path = "~/EnergyMarkR_data/"
)
}
\arguments{
\item{n2ex}{curl handle for the n2ex API}

\item{entso_e}{curl handle for the entso-e transparency platform}

\item{data_path}{Path to folder where data will be stored and read from. Used as default in other
functions in this package. Folder will be created if it doen't exist.}
}
\value{
Saves file "~/gb_elec_creds.Rda" containing specified credentials.
}
\description{
Set-up default credentials and folder/path for use by other functions in this package.
}
\details{
CURL handles sould be in format "username:password".
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
