% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_n2ex_price_vol.R
\name{n2ex_price_vol_download}
\alias{n2ex_price_vol_download}
\title{Download n2ex Price and Volume Data}
\usage{
n2ex_price_vol_download(
  path = paste0(get_credentials()$data_path, "n2ex_price_volume/"),
  h = new_handle(httpauth = 1, userpwd = get_credentials()$n2ex),
  years = 10:(year(Sys.Date()) - 2000)
)
}
\arguments{
\item{path}{Path to folder for n2ex price and volume}

\item{years}{Which years (last two digits, only data from 2000s available) should be downloaded?}

\item{userpwd}{Credentials for the n2ex FTP}
}
\description{
Download or update n2ex price, volume and exchange rate data from the n2ex FTP.
}
\details{
Files are downloaded if a local copy does not exisit. The file corresponding to the last
year in \code{years} is updated.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
