% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ensto_e.R
\name{read_ensto_e}
\alias{read_ensto_e}
\title{Read ENTSO-E Data Files}
\usage{
read_ensto_e(
  data_item = "DayAheadPrices",
  path = paste0(get_credentials()$data_path, "/entsoe_TP/", data_item)
)
}
\arguments{
\item{data_item}{Name of data item}

\item{path}{Path to folder for entso-e data}
}
\value{
A \code{data.table} of data from \code{path}, concatinated and ordered by column \code{DateTime}.
}
\description{
Read ENTSO-E Data Files
}
\details{
Further documentation may be found at \link{https://transparency.entsoe.eu/}.
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
