% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_n2ex_price_vol.R
\name{read_n2ex_price_vol}
\alias{read_n2ex_price_vol}
\title{Read n2EX Price and Volume Data}
\usage{
read_n2ex_price_vol(
  path = paste0(get_credentials()$data_path, "n2ex_price_volume/"),
  years = 15:20
)
}
\arguments{
\item{path}{Path to folder for n2ex price and volume}

\item{years}{Which years (last two digits, only data from 2000s available) should be loaded?}
}
\value{
A \code{data.table} of n2ex price, volume and exchange rate.
}
\description{
Read n2EX Price and Volume Data
}
\author{
Jethro Browell, \email{jethro.browell@strath.ac.uk}
}
